pyproject.toml
[project]
name = "devops-incident-response"
version = "1.0.0"
description = "DevOps Incident Response Environment with AI agent"
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "Jayanth" }]
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "openai",
    "pydantic",
    "python-dotenv",
    "requests",
    "openenv-core"
]

[project.scripts]
server = "server.app:run_server"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["env*", "agent*", "server*"]
